Move ConanIgnoreMatcher to config_install.py
