+++ conans/client/conf/config_installer.py
@@ -1,15 +1,35 @@
 import os
 import shutil
+import fnmatch
 
 from urllib.parse import urlparse, urlsplit
 from contextlib import contextmanager
 
 from conan.api.output import ConanOutput
 from conans.client.downloaders.file_downloader import FileDownloader
-from conans.client.conanignore import ConanIgnoreMatcher
 from conans.errors import ConanException
 from conans.util.files import mkdir, rmdir, remove, unzip, chdir
 from conans.util.runners import detect_runner
+
+
+class ConanIgnoreMatcher:
+    def __init__(self, conanignore_path):
+        self.conanignore_path = os.path.abspath(conanignore_path)
+        self._ignored_entries = {".conanignore"}
+        self._parse_conanignore()
+
+    def _parse_conanignore(self):
+        with open(self.conanignore_path, 'r') as conanignore:
+            for line in conanignore:
+                line_content = line.strip()
+                if line_content != "":
+                    self._ignored_entries.add(line_content)
+
+    def matches(self, path):
+        for ignore_entry in self._ignored_entries:
+            if fnmatch.fnmatch(path, ignore_entry):
+                return True
+        return False
 
 
 def _hide_password(resource):
